\name{humanise.st_result}
\alias{humanise.st_result}
\title{Extract data table from st_result object and add descriptive
labels}
\usage{
\method{humanise}{st_result}(result)
}
\arguments{
  \item{x}{currently defined only for \code{st_result}
  objects}
}
\value{
data frame
}
\description{
Extract data table from st_result object and add
descriptive labels
}

