\docType{package}
\name{seifutoukei}
\alias{seifutoukei}
\alias{seifutoukei-package}
\title{seifutoukei}
\description{
Supports: all of NSTAC's API functionality, except the
server-side query storage and retrieval actions
‘postDataset’ and ‘refDataset’ (and associated
functionality of other actions).
}

