\name{stfind}
\alias{stfind}
\title{Locate tables and other resources in the Japanese official
statistics open data access system}
\usage{
stfind(keywords = NULL, survey.name = NULL, survey.date = NULL)
}
\arguments{
  \item{keywords}{vector of keywords to filter by
  (resources that contain the keywords in metadata will be
  listed)}

  \item{survey.name}{name of a survey to filter by}

  \item{survey.date}{year, year and month (6 digits), or
  interval (year-month1-year-month2)}
}
\description{
Locate tables and other resources in the Japanese official
statistics open data access system
}

