\name{stfind}
\alias{stfind}
\title{Locate tables and other resources in the Japanese official
statistics open data access system}
\usage{
stfind(keywords = NA, survey.name = NA)
}
\arguments{
  \item{keywords}{vector of keywords to filter by
  (resources that contain the keywords in metadata will be
  listed)}

  \item{survey.name}{name of a survey to filter by}
}
\description{
Locate tables and other resources in the Japanese official
statistics open data access system
}

